Split plots.list.test.ts below 500-line limit (warren-8e04)
Decompose the 700-line src/server/handlers/plots.list.test.ts along its
top-level describe seams:
- plots.list.test.ts keeps GET /plots
- plots.needs-attention.test.ts: GET /plots?filter=needs_attention +
  GET /plots/needs-attention/count
- plots.create.test.ts: POST /plots

Shared setup/fixtures hoisted into plots.test-helpers.ts. Remove the
grandfathered budget entry; all resulting files default-pass under the
500-line threshold. Same 21 tests, all green.
