+++ scripts/file-size-budgets.json
@@ -24,7 +24,6 @@
 		"src/server/main/index.ts": 548,
 		"src/server/handlers/projects.settings.test.ts": 685,
 		"src/server/handlers/plot-plan-runs.test.ts": 640,
-		"src/server/handlers/plots.list.test.ts": 700,
 		"src/server/integration.multi-worker.test.ts": 553,
 		"src/server/server.test.ts": 536
 	}
